add mixtral auto policy & move pipeline forward code to modeling folder
